Move parsing from UnitTestDataTree to TestRunning
